Move "FWHeatmapProxyBuilderTemplate" to Fireworks/Calo package
