+++ Fireworks/Calo/plugins/FWHGCalMultiClusterProxyBuilder.cc
@@ -1,5 +1,5 @@
 #include "TEveBoxSet.h"
-#include "Fireworks/Core/interface/FWHeatmapProxyBuilderTemplate.h"
+#include "Fireworks/Calo/interface/FWHeatmapProxyBuilderTemplate.h"
 #include "Fireworks/Core/interface/FWEventItem.h"
 #include "Fireworks/Core/interface/FWGeometry.h"
 #include "Fireworks/Core/interface/BuilderUtils.h"
