+++ Fireworks/SimData/plugins/FWCaloParticleProxyBuilder.cc
@@ -6,21 +6,12 @@
  *
  */
 
-#include "Fireworks/Core/interface/FWHeatmapProxyBuilderTemplate.h"
+#include "Fireworks/Calo/interface/FWHeatmapProxyBuilderTemplate.h"
 #include "Fireworks/Core/interface/Context.h"
-#include "Fireworks/Core/interface/FWEventItem.h"
 #include "Fireworks/Core/interface/FWGeometry.h"
-#include "Fireworks/Core/interface/BuilderUtils.h"
 #include "SimDataFormats/CaloAnalysis/interface/CaloParticle.h"
 #include "SimDataFormats/CaloAnalysis/interface/CaloParticleFwd.h"
 #include "SimDataFormats/CaloAnalysis/interface/SimCluster.h"
-
-#include "FWCore/Common/interface/EventBase.h"
-
-#include "DataFormats/HGCRecHit/interface/HGCRecHitCollections.h"
-#include "DataFormats/DetId/interface/DetId.h"
-
-#include "Fireworks/Core/interface/FWProxyBuilderConfiguration.h"
 
 #include "TEveBoxSet.h"
